# ta-api-master/swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/hotel/book/initiate": {
      "post": {
        "description": "Retrieves available room types, rates, and pricing information for a specific hotel based on check-in/check-out dates and guest requirements.",
        "operationId": "HotelBookController_initiate",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelBookInitiateDto"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Initiate hotel book",
        "tags": [
          "Hotel"
        ]
      }
    },
    "/hotel/book/confirmation": {
      "post": {
        "description": "Retrieves available room types, rates, and pricing information for a specific hotel based on check-in/check-out dates and guest requirements.",
        "operationId": "HotelBookController_bookConfirmation",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelBookConfirmationDto"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Hotel book confirmation",
        "tags": [
          "Hotel"
        ]
      }
    },
    "/hotel/book/booking-details/{bookingRefId}": {
      "get": {
        "description": "Retrieves hotel booking details, including paxes, hotel, contact details, price, and room information, using the provided bookingRefId.",
        "operationId": "HotelBookController_bookingDetails",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          },
          {
            "name": "bookingRefId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get hotel booking details",
        "tags": [
          "Hotel"
        ]
      }
    },
    "/flight/search/start-routing": {
      "post": {
        "operationId": "SearchController_startRouting",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRoutingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Start flight search routing",
        "tags": [
          "Flight"
        ]
      }
    },
    "/flight/search/check-routing": {
      "post": {
        "operationId": "SearchController_collectivePolling",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckRoutingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Check flight search routing",
        "tags": [
          "Flight"
        ]
      }
    },
...
    "/flight/book/initiate": {
      "post": {
        "operationId": "BookController_bookingInitiate",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Book a flight",
        "tags": [
          "Flight"
        ]
      }
    },
    "/flight/book/confirmation": {
      "post": {
        "operationId": "BookController_bookingConfirmation",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookConfirmationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Book a flight",
        "tags": [
          "Flight"
        ]
      }
    },
...
  },
  "info": {
    "title": "API SERVICE",
    "description": "## Travel API Documentation\n\nWelcome to the **API SERVICE** documentation. This guide provides comprehensive guidelines and detailed instructions on how to effectively integrate and utilize the Travel API.\n\n### Overview\nThis documentation includes various sections covering:\n- **Authentication Methods**: Learn how to authenticate using Bearer and Cookie authentication.\n- **Endpoint Descriptions**: Detailed information on each API endpoint.\n- **Usage Examples**: Practical examples to help you get started quickly.\n- **Best Practices**: Tips and recommendations for efficient implementation.\n\n### Purpose\nWhether you are a developer looking to incorporate travel-related functionalities or a business aiming to enhance your travel services, this documentation serves as a valuable resource to guide you through the process.\n\n### Getting Started\nTo get started, ensure you have the necessary authentication tokens and follow the endpoint descriptions to make your first API call.\n\nFor any questions or support, please refer to the contact section at the end of this documentation.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth"
      }
    },
    "schemas": {
...
      "PassengerDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Passenger type (adult, child, infant)",
            "example": "adult"
          },
          "title": {
            "type": "string",
            "description": "Passenger title (Mr, Miss, Mrs, Ms)",
            "example": "Mr"
          },
          "roomId": {
            "type": "number",
            "description": "Passenger roomId (1,2,3)",
            "example": 1
          },
          "age": {
            "type": "number",
            "description": "Passenger age (required for child/infant)",
            "example": 25
          },
          "firstName": {
            "type": "string",
            "description": "First name",
            "example": "John"
          },
          "middleName": {
            "type": "string",
            "description": "middle name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "Last name",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "john.doe@example.com"
          },
          "dob": {
            "type": "string",
            "description": "Date of birth",
            "example": "2002-02-20"
          },
          "dialCode": {
            "type": "string",
            "description": "dial Code",
            "example": "+91"
          },
          "mobileNo": {
            "type": "string",
            "description": "Phone number",
            "example": "+1234567890"
          },
          "nationality": {
            "type": "string",
            "description": "Nationality code",
            "example": "IN"
          },
          "pan": {
            "type": "string",
            "description": "In case of pax is providing PAN then we will consider only pax pan and Parent/Guardian details will be discarded. If pax PAN is incorrect then booking will be failed",
            "example": "CAMPA9865C"
          },
          "passportNumber": {
            "type": "string",
            "description": "In case of pax is providing passport then we will consider only pax passport and Parent/Guardian details will be discarded. If pax passport is incorrect then booking will be failed",
            "example": "M736352"
          },
          "passportIssueDate": {
            "type": "string",
            "description": "Passport issue date a valid date",
            "example": "2025-11-02T00:00:00Z"
          },
          "passportExpDate": {
            "type": "string",
            "description": "Passport expiry date (must be after check-out date)",
            "example": "2025-11-02T00:00:00Z"
          },
          "passportIssuingCountry": {
            "type": "string",
            "description": "Nationality code",
            "example": "IN"
          }
        },
        "required": [
          "type",
          "title",
          "roomId",
          "firstName",
          "lastName"
        ]
      },
...
      "HotelBookInitiateDto": {
        "type": "object",
        "properties": {
          "hotelId": {
            "type": "string",
            "description": "Hotel ID",
            "example": "1863197"
          },
          "supplierCode": {
            "type": "string",
            "description": "Supplier code",
            "example": "TBO"
          },
          "searchReqId": {
            "type": "string",
            "description": "Search request ID from previous search",
            "example": "search_req_12345"
          },
          "rateKey": {
            "type": "string",
            "description": "Unique booking identifier",
            "example": "booking_12345"
          },
          "checkIn": {
            "type": "string",
            "description": "Check-in date (YYYY-MM-DD)",
            "example": "2026-05-15"
          },
          "checkOut": {
            "type": "string",
            "description": "Check-out date (YYYY-MM-DD)",
            "example": "2026-05-16"
          },
          "passengers": {
            "description": "Passenger details",
            "example": [
              {
                "type": "adult",
                "title": "Mr",
                "roomId": 1,
                "firstName": "John",
                "lastName": "Doe",
                "email": "john.doe@example.com",
                "dialCode": "+123",
                "mobileNo": "9627000000",
                "nationality": "US",
                "pan": "CAMPA7654C"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PassengerDto"
            }
          },
          "paymentDetails": {
            "description": "Payment details",
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentDetailsDto"
              }
            ]
          },
          "contactDetails": {
            "description": "Contact details",
            "allOf": [
              {
                "$ref": "#/components/schemas/ContactDetailsDto"
              }
            ]
          }
        },
        "required": [
          "hotelId",
          "supplierCode",
          "searchReqId",
          "rateKey",
          "checkIn",
          "checkOut",
          "passengers",
          "paymentDetails",
          "contactDetails"
        ]
      },
      "HotelBookConfirmationDto": {
        "type": "object",
        "properties": {
          "bookingRefId": {
            "type": "string",
            "description": "Reference booking ID",
            "example": "TA-2343246"
          },
          "searchReqId": {
            "type": "string",
            "description": "Search Request ID",
            "example": "1234567890"
          },
          "paymentLogId": {
            "type": "string",
            "description": "Payment log ID",
            "example": "test-1234567890"
          }
        },
        "required": [
          "bookingRefId",
          "searchReqId",
          "paymentLogId"
        ]
      },
...
      "Paxes": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Type of passenger",
            "example": "ADT",
            "enum": [
              "ADT",
              "CHD",
              "INF"
            ]
          },
          "quantity": {
            "type": "number",
            "description": "Number of passengers of this type",
            "example": 2,
            "minimum": 1
          }
        },
        "required": [
          "type",
          "quantity"
        ]
      },
      "StartRoutingDto": {
        "type": "object",
        "properties": {
          "searchAirLegs": {
            "description": "Array of search air legs",
            "example": [
              {
                "departureDate": "2026-06-15",
                "origin": "NYC",
                "destination": "LAX"
              },
              {
                "departureDate": "2026-06-20",
                "origin": "LAX",
                "destination": "NYC"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchAirLegs"
            }
          },
          "travelPreferences": {
            "description": "Array of travel preferences",
            "example": [
              {
                "maxStopsQuantity": 2,
                "cabinClass": "Economy",
                "airTripType": "roundtrip",
                "nearByAirports": false
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TravelPreference"
            }
          },
          "paxes": {
            "description": "Array of passenger information",
            "example": [
              {
                "type": "ADT",
                "quantity": 2
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Paxes"
            }
          }
        },
        "required": [
          "searchAirLegs",
          "travelPreferences",
          "paxes"
        ]
      },
      "CheckRoutingDto": {
        "type": "object",
        "properties": {
          "searchReqId": {
            "type": "string",
            "description": "Search request ID to check routing status",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          }
        },
        "required": [
          "searchReqId"
        ]
      },
...
      "BookDto": {
        "type": "object",
        "properties": {
          "solutionId": {
            "type": "string",
            "description": "Solution ID for booking",
            "example": "7e3294817ad26fe6a885ca3b57312dbd"
          },
          "searchReqId": {
            "type": "string",
            "description": "Search request ID",
            "example": "efc080e1-3989-427d-abd4-edc7c04d27ab",
            "format": "uuid"
          },
          "trackingId": {
            "type": "string",
            "description": "Tracking ID",
            "example": "bdd3ff73-5ee7-49fc-bafa-34ba0666bfb7"
          },
          "providerCode": {
            "type": "string",
            "description": "Provider code",
            "example": "PK"
          },
          "fareType": {
            "type": "string",
            "description": "Fare type",
            "example": "PUBLIC",
            "enum": [
              "PUBLIC",
              "PRIVATE",
              "CORPORATE"
            ]
          },
          "airTripType": {
            "type": "string",
            "description": "airTripType needed here.",
            "example": "oneway"
          },
          "airlineType": {
            "type": "string",
            "description": "airlineType received in search result, could be LCC, Non-LCC GDS etc.",
            "example": "LCC"
          },
          "passengers": {
            "description": "Passenger information",
            "example": [
              {
                "passengerType": "ADT",
                "gender": "M",
                "passengerDetail": {
                  "title": "Mr",
                  "firstName": "John",
                  "lastName": "Doe"
                },
                "dateOfBirth": "1990-01-15",
                "document": {
                  "documentType": "P",
                  "documentNumber": "A1234567",
                  "expiryDate": "2030-12-31",
                  "country": "US"
                },
                "nationality": "US",
                "mobile": "1234567890",
                "mobileCountryCode": "+1"
              },
              {
                "passengerType": "ADT",
                "gender": "F",
                "passengerDetail": {
                  "title": "Ms",
                  "firstName": "Jane",
                  "lastName": "Doe"
                },
                "dateOfBirth": "1992-03-20",
                "document": {
                  "documentType": "P",
                  "documentNumber": "B7654321",
                  "expiryDate": "2030-12-31",
                  "country": "US"
                },
                "nationality": "US",
                "mobile": "1234567891",
                "mobileCountryCode": "+1"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Passenger"
            }
          },
          "contact": {
            "description": "Contact information",
            "example": {
              "title": "Mr",
              "firstName": "John",
              "lastName": "Doe",
              "gender": "M",
              "email": "john.doe@example.com",
              "mobile": "1234567890",
              "mobileCountryCode": "+1",
              "postalCode": "10001"
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/ContactInfo"
              }
            ]
          },
          "routes": {
            "description": "Route information - array of flight segments for each leg of the journey",
            "example": [
              [
                {
                  "airline": "AA",
                  "departureCode": "LGA",
                  "departureDate": "2026-01-15",
                  "departureTime": "09:29 AM",
                  "arrivalCode": "ORD",
                  "arrivalDate": "2026-01-15",
                  "arrivalTime": "11:10 AM",
                  "bookingCode": "Q",
                  "flightNum": "3234"
                },
                {
                  "airline": "AA",
                  "departureCode": "ORD",
                  "departureDate": "2026-01-15",
                  "departureTime": "02:14 PM",
                  "arrivalCode": "LAX",
                  "arrivalDate": "2026-01-15",
                  "arrivalTime": "04:52 PM",
                  "bookingCode": "Q",
                  "flightNum": "1204"
                }
              ]
            ],
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "gst": {
            "description": "GST details (optional - may be required based on flight/supplier)",
            "example": {
              "gstCompanyAddress": "123 Main Street",
              "gstCompanyContactNumber": "1234567890",
              "gstCompanyName": "ABC Company Ltd.",
              "gstNumber": "27AAAAA0000A1Z5",
              "gstCompanyEmail": "company@example.com"
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/GSTDetails"
              }
            ]
          }
        },
        "required": [
          "solutionId",
          "searchReqId",
          "trackingId",
          "providerCode",
          "fareType",
          "airTripType",
          "airlineType",
          "passengers",
          "contact",
          "routes"
        ]
      },
      "BookConfirmationDto": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string",
            "description": "Booking ID",
            "example": "1234567890"
          },
          "bookingLogId": {
            "type": "string",
            "description": "Booking log ID",
            "example": "1234567890"
          }
        },
        "required": [
          "bookingId",
          "bookingLogId"
        ]
      },
...
    }
  }
}